services/main-server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
requests==2.31.0
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
